Changed position from integer to long, to support up to 100 AU in meters

+++ src/com/oltpbenchmark/benchmarks/galaxy/galaxy-ddl.sql
@@ -23,9 +23,9 @@ CREATE TABLE classes
 CREATE TABLE solar_systems
 (
   solar_system_id integer NOT NULL,
-  max_position_x  integer NOT NULL,
-  max_position_y  integer NOT NULL,
-  max_position_z  integer NOT NULL,
+  max_position_x  bigint  NOT NULL,
+  max_position_y  bigint  NOT NULL,
+  max_position_z  bigint  NOT NULL,
   security_level  integer NOT NULL,
   PRIMARY KEY
   (
@@ -37,9 +37,9 @@ CREATE TABLE solar_systems
 CREATE TABLE ships
 (
   ship_id         integer NOT NULL,
-  position_x      integer NOT NULL,
-  position_y      integer NOT NULL,
-  position_z	  integer NOT NULL,
+  position_x      bigint  NOT NULL,
+  position_y      bigint  NOT NULL,
+  position_z	  bigint  NOT NULL,
   class_id        integer NOT NULL REFERENCES classes (class_id),
   solar_system_id integer NOT NULL REFERENCES solar_systems (solar_system_id),
   health_points   integer NOT NULL,
